backend/requirements.txt
fastapi
uvicorn[standard]
aiogram
sqlalchemy
aiosqlite
python-dotenv
slowapi
apscheduler
pydantic
pydantic-settings
websockets
